requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
attrs==24.2.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
discord.py==2.4.0
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector==2.2.9
numpy==2.1.2
pandas==2.2.3
propcache==0.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
six==1.16.0
tzdata==2024.2
Werkzeug==3.1.3
yarl==1.18.3
